pyproject.toml
[tool.poetry]
name = "quiz-maker"
version = "1.0.0"
description = "api for creating quizzes"
authors = ["Farhad <farhadleon2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
asyncpg = "^0.28.0"
aiosqlite = "^0.19.0"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
pytz = "^2023.3"
uvicorn = "^0.23.2"
certifi = "^2023.7.22"
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
unidecode = "^1.3.7"
redis = "^5.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
ruff = "^0.1.4"
httpx = "^0.25.1"
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = [
    "__init__.py",
    "env.py"
]
line-length = 88
ignore = ['F821']